docker: use docker for "unit testing" as well
We have a couple of unit/integration tests located next to the module.
The problem is that those tests make actual real network requests and
thus depend on the host system.

Build a new container image, named `testing-repo-resource`, that can be
use for running the unit tests instead.

This can be ran with:
$ make test

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>