common: rename sha256sum_from_file to version_from_file
The fact that it's a sha256sum is an implementation detail, and might
actually change in the future.

In fact, since we need fetch a "specific" requested version in _in.py,
we can't use a sha256sum to represent a version.

Prepare for a semantic change of our "version" by renaming the function.

No functional changes.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>