common: version becomes whole content of repo manifest -r -o
In order to implement in_(), we receive a particular version from
concourse. That version, we should be able to fetch it during in_().

With the current representation (a sha256sum), there is no way to
fetch that version.

Use the whole output of repo manifest -r -o instead to represent a
version.
This might be a very long string but it's the only accurate
representation of a "repo state".

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>