common: implement a Version class
Concourse's concept of a resource version is something that our resource
can fetch (in our case over the network).

Implement our version (which is the whole manifest.xml, in one huge
string) in a dedicated Version class.

This is to prepare for the in_() implementation, where we need to
request a specific version to "repo sync"

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>