HACK: common: SourceConfiguration: force gitlab URLs to end with .git
When using an https url from gitlab without suffixing with .git,
repo._Main(['sync']) behaves poorly.
In fact, it generates bogus caracters on stdout, confusing concourse's
version result.

For example, using the following source.url:
https://gitlab.com/baylibre/ti/android/aosp/manifest.git

And run check-resource:
$ fly -t tutorial check-resource -r <resource-name>

generates:
> repo sync has finished successfully.
> Saved manifest to /tmp/tmpc2c992oc/manifest_snapshot
> run check: invalid character 'w' looking for beginning of value
> errored

If we replace this source.url by:
https://gitlab.com/baylibre/ti/android/aosp/manifest.git

we get:
> repo sync has finished successfully.
> Saved manifest to /tmp/tmplxohyvau/manifest_snapshot
> succeeded

So, using `.git` as a suffix does not trigger this (subtle) bug.

As a work-around, bail out early when detecting gitlab http/https
urls that do not have the `.git` suffix.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>