common: add depth argument for optional shallow cloning
As of today, all CHECK/IN operations do shallow cloning for repo init
and repo sync. This saves lots of bandwidth and thus reduces build time
in general.

For CHECK, it always makes sense to do a shallow (depth=1) clone because
we just need to retrieve the version.

For GET, it might be interesting to have a full clone because we might
want to run "git push" operations from a repo that we retrieved via GET.

Add an optional "depth" argument to resource.

NOTE: this changes the default behaviour (which was depth=1 by default)
so pipelines need to be adapted if they want the same fetch speed.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>