check: Fix handling of <remove-project /> element
When `repo-resource` creates a version string (which is a full XML
representation of a manifest), the `<remove-project>` tag is not
taken into account.
Because of this, we generate "false positives" for new Versions,
which could trigger unwanted builds.

For example, with manifest [1], we see that the Version string has
the device/amlogic/yukawa project twice, which is wrong.

Handle the <remove-project /> element to fix this.

[1] https://github.com/makohoek/demo-manifests/blob/main/aosp_remove_yukawa_project.xml
Fixes: https://github.com/makohoek/repo-resource/issues/36
Fixes: c23fc0c2fea1 ("Switch to getRevision code based on git ls-remote")
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>