common: return right revision when git ls-remote
When we execute git ls-remote, the pattern is interpreted as a glob
[1].

Thus we may retrieve the revision of several remote branchs and we
returned the revision of the first branch found.
But the first branch may not be the right branch.

Example with a remote containing these branchs:
a/rev
rev

If we git ls-remote 'rev', we would have returned the revision of
a/rev branch.

The issue is now fixed and we return the revision of the requested
branch.

[1] https://git-scm.com/docs/git-ls-remote/en#Documentation/git-ls-remote.txt-ltpatternsgt82308203

Signed-off-by: Julien Masson <jmasson@baylibre.com>